pyproject.toml
[project]
name = "flax-cnn-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "flax==0.8.5",
    "jax[cuda12]>=0.4.33",
    "orbax-checkpoint>=0.6.4",
    "tensorflow-datasets==4.9.6",
    "tensorflow==2.17.0",
    "typer>=0.12.5",
]
